+++ app/src/main/java/nextstep/shoppingcart/ui/shoppingdetail/component/ShoppingDetailProductTitle.kt
@@ -8,8 +8,8 @@ import androidx.compose.ui.graphics.Color
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
 import androidx.compose.ui.unit.sp
+import nextstep.shoppingcart.data.dummyProducts
 import nextstep.shoppingcart.ui.shoppinglist.model.Product
-import nextstep.shoppingcart.ui.shoppinglist.model.dummyProducts
 import nextstep.shoppingcart.ui.theme.RobotoBold
 
 @Composable
