+++ app/src/main/java/nextstep/shoppingcart/ui/shoppinglist/component/ShoppingListItem.kt
@@ -18,9 +18,9 @@ import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
 import androidx.compose.ui.unit.sp
 import nextstep.shoppingcart.R
+import nextstep.shoppingcart.data.dummyProducts
 import nextstep.shoppingcart.ui.component.ShoppingProductImage
 import nextstep.shoppingcart.ui.shoppinglist.model.Product
-import nextstep.shoppingcart.ui.shoppinglist.model.dummyProducts
 import nextstep.shoppingcart.ui.theme.RobotoBold
 import nextstep.shoppingcart.ui.theme.RobotoRegular
